extract ContributionField ans Contributor from AddFieldDetail
